requirements.txt
Django==1.7.7
boto==2.38.0
dj-database-url==0.3.0
django-cors-headers==1.0.0
django-custom-user==0.5
djangorestframework==3.1.0
djangorestframework-camel-case==0.2.0
gunicorn==19.3.0
psycopg2==2.6
wsgiref==0.1.2
